+++ examples/todos-ts/src/stores/index.ts
@@ -1,8 +1,6 @@
-import Icestore from '@ice/store';
+import createIcestore from '@ice/store';
 import logger from '@ice/store-logger';
 import todos from './todos';
-
-const icestore = new Icestore();
 
 const middlewares: ((ctx, next) => {})[] = [];
 
@@ -10,10 +8,4 @@ if (process.env.NODE_ENV !== 'production') {
   middlewares.push(logger);
 }
 
-icestore.applyMiddleware(middlewares);
-
-const stores = icestore.registerStores({
-  todos,
-});
-
-export default stores;
+export default createIcestore(todos, middlewares);
