+++ examples/todos/src/stores/index.js
@@ -1,19 +1,10 @@
-import Icestore from '@ice/store';
+import createIcestore from '@ice/store';
 import logger from '@ice/store-logger';
 import todos from './todos';
 
-const icestore = new Icestore('todos');
-
 const middlewares = [];
-
 if (process.env.NODE_ENV !== 'production') {
   middlewares.push(logger);
 }
 
-icestore.applyMiddleware(middlewares);
-
-const stores = icestore.registerStores({
-  todos,
-});
-
-export default stores;
+export default createIcestore(todos, middlewares);
